backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
requests>=2.31.0
google-genai>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pillow>=10.0.0
